+++ src/Components/Sidebar/FilterHeader/FilterHeader.jsx
@@ -1,4 +1,4 @@
-import { useFilterReducerContext } from "../../../context/useFilterReducer.context";
+import { useFilterReducerContext } from "../../../context/FilterReducer.context";
 
 export const FilterHeader = () => {
   const { sidebar, dispatch } = useFilterReducerContext();
