+++ src/Components/Sidebar/SortFilter/SortFilter.jsx
@@ -1,4 +1,4 @@
-import { useFilterReducerContext } from "../../../context/useFilterReducer.context";
+import { useFilterReducerContext } from "../../../context/FilterReducer.context";
 
 export const SortFilter = () => {
   const { sort, dispatch } = useFilterReducerContext();
