+++ src/Pages/ProductPage/ProductComponents/SectionContainer/SectionContainer.jsx
@@ -1,4 +1,4 @@
-import { useFilterReducerContext } from "../../../../context/useFilterReducer.context";
+import { useFilterReducerContext } from "../../../../context/FilterReducer.context";
 import "./SectionContainer.css";
 
 export const SectionContainer = ({ children }) => {
