+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom";
 import App from "./App";
 import { makeServer } from "./server";
 import { BrowserRouter as Router } from "react-router-dom";
-import { FilterReducerProvider } from "./context/useFilterReducer.context";
+import { FilterReducerProvider } from "./context/FilterReducer.context";
 import { useSignUp } from "./Hook/useSignUp";
 
 // Call make Server
